projects/image-organizer/requirements.txt
streamlit
pillow
exifread
pandas
torch
torchvision
transformers
tqdm
